# web/src/lib/schemas/openapi-v3.yml
openapi: 3.1.0
info:
  title: Talaria V3 entity projections
  version: 1.0.0
paths:
...
  /api/v1/interactions:
    post:
      summary: Create or upsert a stateful Talaria interaction
      description: |
        Session cookie required. `user_id` is taken from the session, never the body.
        Idempotent for non-stance actions: first insert returns 201, duplicate returns 200 with the same row.
        Claim stances (support/dispute/uncertain) replace transactionally so exactly one stance remains.
        No Intuition fields (tx_hash, term_id, intuition_status).
        source target v1 = corpus_documents.id (Scholar corpus document).
        route is not supported (Visit has no durable itinerary UUID; TODO PR8).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InteractionCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InteractionEnvelope'
        '200':
          description: Existing row returned (idempotent upsert)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InteractionEnvelope'
        '400': {$ref: '#/components/responses/StructuredError'}
        '401': {$ref: '#/components/responses/StructuredError'}
        '404': {$ref: '#/components/responses/StructuredError'}
        '500': {$ref: '#/components/responses/StructuredError'}
  /api/v1/interactions/{id}:
    delete:
      summary: Delete an interaction owned by the session user
      description: Foreign ids return 404 (no existence leak).
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string, format: uuid}
      responses:
        '204': {description: Deleted}
        '401': {$ref: '#/components/responses/StructuredError'}
        '404': {$ref: '#/components/responses/StructuredError'}
        '500': {$ref: '#/components/responses/StructuredError'}
  /api/v1/me/interactions:
    get:
      summary: List the session user's interactions (cursor pagination)
      parameters:
        - name: action
          in: query
          schema: {$ref: '#/components/schemas/InteractionAction'}
        - name: target_type
          in: query
          schema: {$ref: '#/components/schemas/InteractionTargetType'}
        - name: cursor
          in: query
          schema: {type: string}
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 100, default: 50}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                required: [items, next_cursor]
                properties:
                  items:
                    type: array
                    items: {$ref: '#/components/schemas/Interaction'}
                  next_cursor: {type: [string, 'null']}
        '401': {$ref: '#/components/responses/StructuredError'}
        '400': {$ref: '#/components/responses/StructuredError'}
  /api/v1/interactions/summary:
    get:
      summary: Batch public interaction counts (max 100 ids, one grouped query)
      description: |
        Public, no auth required. `counts` include only visibility=public.
        `mine` is empty for anonymous callers; authenticated callers see their rows including private.
      parameters:
        - name: target_type
          in: query
          required: true
          schema: {$ref: '#/components/schemas/InteractionTargetType'}
        - name: target_ids
          in: query
          required: true
          description: Comma-separated UUIDs, max 100
          schema: {type: string}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                required: [summaries]
                properties:
                  summaries:
                    type: array
                    items: {$ref: '#/components/schemas/InteractionSummary'}
        '400': {$ref: '#/components/responses/StructuredError'}
...
components:
...
  responses:
...
    StructuredError:
      description: Code + message, never raw SQL
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error:
                type: object
                required: [code, message]
                properties:
                  code:
                    type: string
                    enum:
                      - unauthenticated
                      - invalid_action
                      - invalid_target_type
                      - invalid_action_for_target
                      - target_not_found
                      - interaction_not_found
                      - invalid_visibility
                      - batch_too_large
                      - database_error
                      - invalid_address
                      - conflict
                      - expired_challenge
                      - invalid_signature
                  message: {type: string}
  schemas:
    InteractionAction:
      type: string
      enum:
        - interest
        - follow
        - save
        - support
        - dispute
        - uncertain
        - useful
        - credible
        - not_credible
        - want_to_visit
        - visited
    InteractionTargetType:
      type: string
      enum: [person, claim, source, place, event]
      description: route is omitted until Visit has a durable UUID (PR8)
    InteractionVisibility:
      type: string
      enum: [private, public]
    InteractionCreate:
      type: object
      required: [action, target_type, target_id]
      properties:
        action: {$ref: '#/components/schemas/InteractionAction'}
        target_type: {$ref: '#/components/schemas/InteractionTargetType'}
        target_id: {type: string, format: uuid}
        visibility: {$ref: '#/components/schemas/InteractionVisibility'}
      description: |
        Visibility defaults: save/follow/interest/visited/want_to_visit=private;
        support/dispute/uncertain/useful/credible/not_credible=public.
        Never send user_id or wallet_address.
    Interaction:
      type: object
      required: [id, action, target_type, target_id, visibility, created_at, updated_at]
      properties:
        id: {type: string, format: uuid}
        action: {$ref: '#/components/schemas/InteractionAction'}
        target_type: {$ref: '#/components/schemas/InteractionTargetType'}
        target_id: {type: string, format: uuid}
        visibility: {$ref: '#/components/schemas/InteractionVisibility'}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    InteractionEnvelope:
      type: object
      required: [interaction]
      properties:
        interaction: {$ref: '#/components/schemas/Interaction'}
    InteractionSummary:
      type: object
      required: [target_id, counts, comment_count, mine]
      properties:
        target_id: {type: string, format: uuid}
        counts:
          type: object
          additionalProperties: {type: integer}
        comment_count: {type: integer}
        argument_count: {type: integer}
        source_count: {type: integer}
        mine:
          type: array
          items: {$ref: '#/components/schemas/InteractionAction'}
...